move sentimentType to consts
